requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
email-validator
alembic==1.13.2
python-jose[cryptography]
python-multipart
pytest
pytest-asyncio
httpx
fastapi-mail==1.4.1
email-validator==2.2.0
SQLAlchemy==2.0.31
psycopg[binary]==3.2.1
pydantic-settings==2.5.2
python-dotenv==1.0.1



